seperate multiple label logic from existing deepedit app/code
