+++ monailabel/interfaces/tasks/infer.py
@@ -240,7 +240,6 @@ class InferTask:
 
         start = time.time()
         result_file_name, result_json = self.writer(data)
-        # _, _ = self.writer(data) # Keep output in /tmp folder
         latency_write = time.time() - start
 
         latency_total = time.time() - begin
