+++ sample-apps/deepedit_multilabel/lib/transforms.py
