+++ sample-apps/deepedit_multilabel/main.py
@@ -49,6 +49,7 @@ class MyApp(MONAILabelApp):
             "liver": 6,
             "background": 0,
         }
+        self.labels = ["spleen", "right_kidney", "left_kidney", "liver"]
 
         network_params = {
             "spatial_dims": 3,
@@ -95,7 +96,7 @@ class MyApp(MONAILabelApp):
         self.planner = HeuristicPlanner(spatial_size=spatial_size, target_spacing=target_spacing)
 
         use_pretrained_model = strtobool(conf.get("use_pretrained_model", "false"))
-        pretrained_model_uri = conf.get("pretrained_model_path", f"{self.PRE_TRAINED_PATH}/deepedit_spleen.pt")
+        pretrained_model_uri = conf.get("pretrained_model_path", f"{self.PRE_TRAINED_PATH}/deepedit_multilabel.pt")
 
         # Path to pretrained weights
         if use_pretrained_model:
@@ -154,7 +155,7 @@ class MyApp(MONAILabelApp):
                 load_path=self.pretrained_model,
                 publish_path=self.final_model,
                 config={"pretrained": strtobool(self.conf.get("use_pretrained_model", "true"))},
-                label_names=self.label_names,
+                label_names=self.labels,
                 debug_mode=False,
             )
         }
