+++ sample-apps/deepedit_multilabel/model/.gitignore
@@ -1,0 +1,1 @@
+*
